Integrate the nv-compact active-DOF solver into the existing modules
Folds the experimental active-DOF compaction feature out of the standalone nvmax.py module and into the modules that already own the equivalent functionality, mirroring how the island solver is structured. The active-set bookkeeping (update_active_dofs and its map-building kernels) moves to island.py next to compute_island_mapping, and the compacted smooth and constrained solves plus their gather/scatter kernels move to solver.py next to the island gather/scatter.

The compacted workspace was previously held in an NvCompactContext cached on a monkey-patched d._nvcompact_ctx attribute. It now lives as c-prefixed Data fields (cM, cJ, cqacc, ...) parallel to the island i-prefixed fields, allocated in io.py and gated on whether the user passed nvmax. The compacted dense matrix and DOF vectors stay separate from the full-size d.M / d.qacc because they have a different index space (the packed active set) and lifetime (the full arrays are still needed for scatter-back, the integrators, and warmstart). A TODO in io._allocate_compact_arrays notes that once compaction replaces the island solver and the whole forward pipeline runs compacted, these shadows can collapse into nvmax-sized primary fields.

nvmax_test.py is renamed to compact_test.py; the solve-equivalence tests now build Data directly with an explicit nvmax since test_data.fixture does not expose it.
