--- mujoco_warp/_src/nvmax.py
@@ -1,576 +1,0 @@
-# Copyright 2026 The Newton Developers
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Experimental: compact active DOFs into nvmax-sized arrays for factor/solve.
-
-The active set is tracked per kinematic tree (the unit of coupling in the mass
-matrix). A tree becomes active through an actuator (seeded at make_data), an
-applied force, or a contact, and stays active (monotonic - no deactivation yet).
-Each step the active trees' DOFs are packed into a contiguous [0, ncdof) range
-so the dense factor/solver can run at size nvmax instead of nv.
-"""
-
-import dataclasses
-
-import warp as wp
-
-from mujoco_warp._src import solver
-from mujoco_warp._src import support
-from mujoco_warp._src import types
-from mujoco_warp._src.block_cholesky import create_blocked_cholesky_factorize_solve_func
-from mujoco_warp._src.warp_util import cache_kernel
-from mujoco_warp._src.warp_util import event_scope
-
-_TILE_SIZE = types.TILE_SIZE_JTDAJ_DENSE
-
-
-def _round_up(x: int, multiple: int) -> int:
-  return ((x + multiple - 1) // multiple) * multiple
-
-
-@wp.kernel
-def _reset_compact_maps(
-  # Model:
-  nv: int,
-  # Data in:
-  nvmax_in: int,
-  # Data out:
-  dof_cdof_out: wp.array2d[int],
-  cdof_dof_out: wp.array2d[int],
-):
-  worldid, idx = wp.tid()
-  if idx < nv:
-    dof_cdof_out[worldid, idx] = -1
-  if idx < nvmax_in:
-    cdof_dof_out[worldid, idx] = -1
-
-
-@wp.kernel
-def _compact_dofs(
-  # Model:
-  ntree: int,
-  tree_dofadr: wp.array[int],
-  tree_dofnum: wp.array[int],
-  # Data in:
-  tree_awake_in: wp.array2d[int],
-  nvmax_in: int,
-  # Data out:
-  ncdof_out: wp.array[int],
-  dof_cdof_out: wp.array2d[int],
-  cdof_dof_out: wp.array2d[int],
-):
-  worldid = wp.tid()
-  count = int(0)
-  for t in range(ntree):
-    if tree_awake_in[worldid, t] == 1:
-      adr = tree_dofadr[t]
-      num = tree_dofnum[t]
-      for j in range(num):
-        dof = adr + j
-        if count < nvmax_in:
-          dof_cdof_out[worldid, dof] = count
-          cdof_dof_out[worldid, count] = dof
-        count += 1
-
-  if count > nvmax_in:
-    wp.printf(
-      "nvmax overflow: world %d needs %d active DOFs but nvmax = %d (behavior undefined)\n",
-      worldid,
-      count,
-      nvmax_in,
-    )
-    ncdof_out[worldid] = nvmax_in
-  else:
-    ncdof_out[worldid] = count
-
-
-@event_scope
-def update_active_dofs(m: types.Model, d: types.Data):
-  """Rebuild the compaction maps from tree_awake."""
-  wp.launch(
-    _reset_compact_maps,
-    dim=(d.nworld, m.nv),
-    inputs=[m.nv, d.nvmax],
-    outputs=[d.dof_cdof, d.cdof_dof],
-  )
-  wp.launch(
-    _compact_dofs,
-    dim=(d.nworld,),
-    inputs=[m.ntree, m.tree_dofadr, m.tree_dofnum, d.tree_awake, d.nvmax],
-    outputs=[d.ncdof, d.dof_cdof, d.cdof_dof],
-  )
-
-
-@dataclasses.dataclass
-class NvCompactContext:
-  """Workspace for the compacted dense factor/solve.
-
-  DOF-space arrays are sized by nvmax_pad (nvmax rounded up to the dense tile size)
-  so the blocked Cholesky never reads out of bounds on its partial block.
-  """
-
-  nvmax_pad: int
-  # convergence tolerances rescaled by nv_full/nvmax_pad so the solver's nv-normalized
-  # done-test (solve.py: value/(meaninertia*nv)) matches the full-model baseline.
-  tol_c: wp.array
-  ls_tol_c: wp.array
-  # compacted dof-pair indices for the elliptic JTCJ Hessian (over nvmax_pad, not global nv)
-  dof_tri_row_c: wp.array
-  dof_tri_col_c: wp.array
-  # smooth-solve workspace (Stage 2a)
-  M_c: wp.array3d[float]  # (nworld, nvmax_pad, nvmax_pad) compacted dense inertia
-  qLD_c: wp.array3d[float]  # (nworld, nvmax_pad, nvmax_pad) upper Cholesky factor
-  rhs_c: wp.array3d[float]  # (nworld, nvmax_pad, 1) compacted right-hand side
-  x_c: wp.array3d[float]  # (nworld, nvmax_pad, 1) compacted solution
-  # constrained-solve workspace (Stage 2b): compacted views fed to the dense Newton solver
-  J_c: wp.array3d[float]  # (nworld, njmax_pad, nvmax_pad) compacted dense constraint Jacobian
-  Ma_c: wp.array2d[float]  # (nworld, nvmax_pad) M @ qacc workspace
-  qfrc_smooth_c: wp.array2d[float]
-  qacc_smooth_c: wp.array2d[float]
-  qacc_warmstart_c: wp.array2d[float]
-  qacc_c: wp.array2d[float]
-  qfrc_constraint_c: wp.array2d[float]
-
-
-@wp.kernel
-def _scale_tolerances(
-  # In:
-  tolerance: wp.array[float],
-  ls_tolerance: wp.array[float],
-  scale: float,
-  # Out:
-  tol_out: wp.array[float],
-  ls_tol_out: wp.array[float],
-):
-  i = wp.tid()
-  tol_out[i] = tolerance[i] * scale
-  ls_tol_out[i] = ls_tolerance[i] * scale
-
-
-@wp.kernel
-def _fill_dof_pairs(n: int, row_out: wp.array[int], col_out: wp.array[int]):
-  # Enumerate all (i, j) DOF pairs of the nvmax_pad-wide compacted Hessian. The elliptic
-  # JTCJ correction indexes the Hessian via these; the global dof_tri (triu over full nv)
-  # would write out of bounds. Pairs with a zero J_c column contribute nothing.
-  i, j = wp.tid()
-  row_out[i * n + j] = i
-  col_out[i * n + j] = j
-
-
-def get_context(m: types.Model, d: types.Data) -> NvCompactContext:
-  """Return a cached compaction context for d, creating it on first use.
-
-  Cached as a dynamic attribute so it is allocated once (outside any graph capture)
-  and reused across steps.
-  """
-  ctx = getattr(d, "_nvcompact_ctx", None)
-  if ctx is None or ctx.nvmax_pad != _round_up(max(d.nvmax, 1), _TILE_SIZE):
-    ctx = create_nvcompact_context(m, d)
-    d._nvcompact_ctx = ctx
-  return ctx
-
-
-def create_nvcompact_context(m: types.Model, d: types.Data) -> NvCompactContext:
-  nvmax_pad = _round_up(max(d.nvmax, 1), _TILE_SIZE)
-  nworld = d.nworld
-
-  # match baseline's nv-normalized convergence test: the compact solve passes nv=nvmax_pad
-  # (< full nv) to the rescale, making it stricter; scale tolerance up to compensate.
-  # Done on-device (no host sync) so context creation is safe inside CUDA graph capture.
-  tol_scale = float(m.nv) / float(nvmax_pad)
-  tol_c = wp.empty_like(m.opt.tolerance)
-  ls_tol_c = wp.empty_like(m.opt.ls_tolerance)
-  wp.launch(
-    _scale_tolerances,
-    dim=tol_c.shape[0],
-    inputs=[m.opt.tolerance, m.opt.ls_tolerance, tol_scale],
-    outputs=[tol_c, ls_tol_c],
-  )
-
-  # compacted dof-pair indices for the elliptic JTCJ Hessian (built on-device, capture-safe)
-  npair = nvmax_pad * nvmax_pad
-  dof_tri_row_c = wp.empty(npair, dtype=int)
-  dof_tri_col_c = wp.empty(npair, dtype=int)
-  wp.launch(_fill_dof_pairs, dim=(nvmax_pad, nvmax_pad), inputs=[nvmax_pad], outputs=[dof_tri_row_c, dof_tri_col_c])
-
-  return NvCompactContext(
-    nvmax_pad=nvmax_pad,
-    tol_c=tol_c,
-    ls_tol_c=ls_tol_c,
-    dof_tri_row_c=dof_tri_row_c,
-    dof_tri_col_c=dof_tri_col_c,
-    M_c=wp.empty((nworld, nvmax_pad, nvmax_pad), dtype=float),
-    qLD_c=wp.empty((nworld, nvmax_pad, nvmax_pad), dtype=float),
-    rhs_c=wp.empty((nworld, nvmax_pad, 1), dtype=float),
-    x_c=wp.empty((nworld, nvmax_pad, 1), dtype=float),
-    J_c=wp.empty((nworld, d.njmax_pad, nvmax_pad), dtype=float),
-    Ma_c=wp.empty((nworld, nvmax_pad), dtype=float),
-    qfrc_smooth_c=wp.empty((nworld, nvmax_pad), dtype=float),
-    qacc_smooth_c=wp.empty((nworld, nvmax_pad), dtype=float),
-    qacc_warmstart_c=wp.empty((nworld, nvmax_pad), dtype=float),
-    qacc_c=wp.empty((nworld, nvmax_pad), dtype=float),
-    qfrc_constraint_c=wp.empty((nworld, nvmax_pad), dtype=float),
-  )
-
-
-@wp.kernel
-def _init_compact_inertia(
-  # Data in:
-  ncdof_in: wp.array[int],
-  # Out:
-  M_c_out: wp.array3d[float],
-):
-  worldid, i, j = wp.tid()
-  val = 0.0
-  if i == j and i >= ncdof_in[worldid]:
-    val = 1.0
-  M_c_out[worldid, i, j] = val
-
-
-@wp.kernel
-def _gather_M_sparse(
-  # Model:
-  M_rownnz: wp.array[int],
-  M_rowadr: wp.array[int],
-  M_colind: wp.array[int],
-  # Data in:
-  M_in: wp.array3d[float],
-  dof_cdof_in: wp.array2d[int],
-  # Out:
-  M_c_out: wp.array3d[float],
-):
-  worldid, i = wp.tid()
-  ci = dof_cdof_in[worldid, i]
-  if ci < 0:
-    return
-  rowadr = M_rowadr[i]
-  for k in range(M_rownnz[i]):
-    adr = rowadr + k
-    cj = dof_cdof_in[worldid, M_colind[adr]]
-    if cj >= 0:
-      val = M_in[worldid, 0, adr]
-      M_c_out[worldid, ci, cj] = val
-      M_c_out[worldid, cj, ci] = val
-
-
-@wp.kernel
-def _gather_M_dense(
-  # Data in:
-  M_in: wp.array3d[float],
-  dof_cdof_in: wp.array2d[int],
-  # Out:
-  M_c_out: wp.array3d[float],
-):
-  worldid, i = wp.tid()
-  ci = dof_cdof_in[worldid, i]
-  if ci < 0:
-    return
-  nv = dof_cdof_in.shape[1]
-  for j in range(nv):
-    cj = dof_cdof_in[worldid, j]
-    if cj >= 0:
-      val = M_in[worldid, i, j]
-      M_c_out[worldid, ci, cj] = val
-
-
-@wp.kernel
-def _gather_rhs_compact(
-  # Data in:
-  cdof_dof_in: wp.array2d[int],
-  # In:
-  vec_in: wp.array2d[float],
-  # Out:
-  rhs_out: wp.array3d[float],
-):
-  worldid, ci = wp.tid()
-  dof = cdof_dof_in[worldid, ci]
-  if dof >= 0:
-    rhs_out[worldid, ci, 0] = vec_in[worldid, dof]
-  else:
-    rhs_out[worldid, ci, 0] = 0.0
-
-
-@wp.kernel
-def _scatter_solution(
-  # Data in:
-  dof_cdof_in: wp.array2d[int],
-  # In:
-  x_in: wp.array3d[float],
-  # Out:
-  vec_out: wp.array2d[float],
-):
-  worldid, i = wp.tid()
-  ci = dof_cdof_in[worldid, i]
-  if ci >= 0:
-    vec_out[worldid, i] = x_in[worldid, ci, 0]
-  else:
-    vec_out[worldid, i] = 0.0  # frozen inactive DOF
-
-
-@cache_kernel
-def _blocked_factor_solve(matrix_size: int):
-  @wp.kernel(module="unique", enable_backward=False)
-  def kernel(
-    # Data in:
-    M_in: wp.array3d[float],
-    # In:
-    rhs_in: wp.array3d[float],
-    msize: int,
-    # Out:
-    U_out: wp.array3d[float],
-    x_out: wp.array3d[float],
-  ):
-    worldid = wp.tid()
-    wp.static(create_blocked_cholesky_factorize_solve_func(_TILE_SIZE, matrix_size))(
-      M_in[worldid], rhs_in[worldid], msize, U_out[worldid], x_out[worldid]
-    )
-
-  return kernel
-
-
-@event_scope
-def smooth_solve_compact(m: types.Model, d: types.Data, ctx: NvCompactContext):
-  """Compacted equivalent of solve_m: qacc_smooth[active] = M_c^-1 qfrc_smooth[active].
-
-  Inactive DOFs are frozen (qacc_smooth set to 0). Reads the sparse Model inertia.
-  """
-  wp.launch(
-    _init_compact_inertia,
-    dim=(d.nworld, ctx.nvmax_pad, ctx.nvmax_pad),
-    inputs=[d.ncdof],
-    outputs=[ctx.M_c],
-  )
-  wp.launch(
-    _gather_M_sparse,
-    dim=(d.nworld, m.nv),
-    inputs=[m.M_rownnz, m.M_rowadr, m.M_colind, d.M, d.dof_cdof],
-    outputs=[ctx.M_c],
-  )
-  wp.launch(
-    _gather_rhs_compact,
-    dim=(d.nworld, ctx.nvmax_pad),
-    inputs=[d.cdof_dof, d.qfrc_smooth],
-    outputs=[ctx.rhs_c],
-  )
-  wp.launch_tiled(
-    _blocked_factor_solve(ctx.nvmax_pad),
-    dim=d.nworld,
-    inputs=[ctx.M_c, ctx.rhs_c, ctx.nvmax_pad],
-    outputs=[ctx.qLD_c, ctx.x_c],
-    block_dim=m.block_dim.update_gradient_cholesky_blocked,
-  )
-  wp.launch(_scatter_solution, dim=(d.nworld, m.nv), inputs=[d.dof_cdof, ctx.x_c], outputs=[d.qacc_smooth])
-
-
-@wp.kernel
-def _gather_dof_vecs_compact(
-  # Data in:
-  qacc_warmstart_in: wp.array2d[float],
-  qfrc_smooth_in: wp.array2d[float],
-  qacc_smooth_in: wp.array2d[float],
-  cdof_dof_in: wp.array2d[int],
-  # Out:
-  qfrc_smooth_c_out: wp.array2d[float],
-  qacc_smooth_c_out: wp.array2d[float],
-  qacc_warmstart_c_out: wp.array2d[float],
-):
-  worldid, ci = wp.tid()
-  dof = cdof_dof_in[worldid, ci]
-  if dof >= 0:
-    qfrc_smooth_c_out[worldid, ci] = qfrc_smooth_in[worldid, dof]
-    qacc_smooth_c_out[worldid, ci] = qacc_smooth_in[worldid, dof]
-    qacc_warmstart_c_out[worldid, ci] = qacc_warmstart_in[worldid, dof]
-  else:
-    qfrc_smooth_c_out[worldid, ci] = 0.0
-    qacc_smooth_c_out[worldid, ci] = 0.0
-    qacc_warmstart_c_out[worldid, ci] = 0.0
-
-
-@wp.kernel
-def _scatter_dof_vecs(
-  # Data in:
-  dof_cdof_in: wp.array2d[int],
-  # In:
-  qacc_c_in: wp.array2d[float],
-  qfrc_constraint_c_in: wp.array2d[float],
-  # Data out:
-  qacc_out: wp.array2d[float],
-  qfrc_constraint_out: wp.array2d[float],
-):
-  worldid, i = wp.tid()
-  ci = dof_cdof_in[worldid, i]
-  if ci >= 0:
-    qacc_out[worldid, i] = qacc_c_in[worldid, ci]
-    qfrc_constraint_out[worldid, i] = qfrc_constraint_c_in[worldid, ci]
-  else:
-    qacc_out[worldid, i] = 0.0
-    qfrc_constraint_out[worldid, i] = 0.0
-
-
-@wp.kernel
-def _gather_J_sparse(
-  # Data in:
-  nefc_in: wp.array[int],
-  dof_cdof_in: wp.array2d[int],
-  # In:
-  J_rownnz_in: wp.array2d[int],
-  J_rowadr_in: wp.array2d[int],
-  J_colind_in: wp.array3d[int],
-  J_in: wp.array3d[float],
-  # Out:
-  J_c_out: wp.array3d[float],
-):
-  worldid, efcid = wp.tid()
-  if efcid >= nefc_in[worldid]:
-    return
-  rowadr = J_rowadr_in[worldid, efcid]
-  for k in range(J_rownnz_in[worldid, efcid]):
-    adr = rowadr + k
-    cj = dof_cdof_in[worldid, J_colind_in[worldid, 0, adr]]
-    if cj >= 0:
-      J_c_out[worldid, efcid, cj] = J_in[worldid, 0, adr]
-
-
-@wp.kernel
-def _gather_J_dense(
-  # Data in:
-  nefc_in: wp.array[int],
-  dof_cdof_in: wp.array2d[int],
-  # In:
-  J_in: wp.array3d[float],
-  # Out:
-  J_c_out: wp.array3d[float],
-):
-  worldid, efcid = wp.tid()
-  if efcid >= nefc_in[worldid]:
-    return
-  nv = dof_cdof_in.shape[1]
-  for j in range(nv):
-    cj = dof_cdof_in[worldid, j]
-    if cj >= 0:
-      J_c_out[worldid, efcid, cj] = J_in[worldid, efcid, j]
-
-
-@event_scope
-def solve_compact(m: types.Model, d: types.Data, ctx: NvCompactContext):
-  """Run the dense Newton constraint solver in compacted DOF space.
-
-  Gathers the active-DOF inertia, constraint Jacobian, and smooth/warmstart vectors
-  into nvmax_pad-sized dense workspaces, runs the stock dense Newton solver on a
-  shallow-replaced (m, d) at nvmax_pad, then scatters qacc/qfrc_constraint back.
-  Inactive DOFs are frozen to 0. Reads the sparse Model inertia and constraint J.
-  """
-  nvp = ctx.nvmax_pad
-
-  _compact_gather(m, d, ctx)
-
-  # shallow-replace (m, d) so the stock dense Newton solver runs at nvmax_pad.
-  # Keep graph-conditional early-exit on CUDA (matches baseline: stops at convergence
-  # instead of running all iterations); fall back to the plain loop on CPU.
-  gc = m.opt.graph_conditional and wp.get_device().is_cuda
-  opt2 = dataclasses.replace(m.opt, graph_conditional=gc, tolerance=ctx.tol_c, ls_tolerance=ctx.ls_tol_c)
-  m2 = dataclasses.replace(
-    m, opt=opt2, nv=nvp, nv_pad=nvp, is_sparse=False, dof_tri_row=ctx.dof_tri_row_c, dof_tri_col=ctx.dof_tri_col_c
-  )
-  efc2 = dataclasses.replace(d.efc, J=ctx.J_c, Ma=ctx.Ma_c)
-  d2 = dataclasses.replace(
-    d,
-    M=ctx.M_c,
-    qfrc_smooth=ctx.qfrc_smooth_c,
-    qacc_smooth=ctx.qacc_smooth_c,
-    qacc_warmstart=ctx.qacc_warmstart_c,
-    qacc=ctx.qacc_c,
-    qfrc_constraint=ctx.qfrc_constraint_c,
-    efc=efc2,
-  )
-
-  sctx = solver._create_solver_context(m2, d2)
-  solver._solve(m2, d2, sctx)
-
-  _compact_scatter(m, d, ctx)
-
-
-@event_scope
-def _compact_gather(m: types.Model, d: types.Data, ctx: NvCompactContext):
-  nvp = ctx.nvmax_pad
-  # gather compacted dense inertia (identity-padded tail)
-  wp.launch(
-    _init_compact_inertia,
-    dim=(d.nworld, nvp, nvp),
-    inputs=[d.ncdof],
-    outputs=[ctx.M_c],
-  )
-
-  if m.is_sparse:
-    wp.launch(
-      _gather_M_sparse,
-      dim=(d.nworld, m.nv),
-      inputs=[m.M_rownnz, m.M_rowadr, m.M_colind, d.M, d.dof_cdof],
-      outputs=[ctx.M_c],
-    )
-  else:
-    wp.launch(
-      _gather_M_dense,
-      dim=(d.nworld, m.nv),
-      inputs=[d.M, d.dof_cdof],
-      outputs=[ctx.M_c],
-    )
-  # gather compacted dense constraint Jacobian (active columns only)
-  ctx.J_c.zero_()
-  if m.is_sparse:
-    wp.launch(
-      _gather_J_sparse,
-      dim=(d.nworld, d.njmax),
-      inputs=[d.nefc, d.dof_cdof, d.efc.J_rownnz, d.efc.J_rowadr, d.efc.J_colind, d.efc.J],
-      outputs=[ctx.J_c],
-    )
-  else:
-    wp.launch(
-      _gather_J_dense,
-      dim=(d.nworld, d.njmax),
-      inputs=[d.nefc, d.dof_cdof, d.efc.J],
-      outputs=[ctx.J_c],
-    )
-  # gather compacted DOF-space vectors in a single launch
-  wp.launch(
-    _gather_dof_vecs_compact,
-    dim=(d.nworld, nvp),
-    inputs=[
-      d.qacc_warmstart,
-      d.qfrc_smooth,
-      d.qacc_smooth,
-      d.cdof_dof,
-    ],
-    outputs=[
-      ctx.qfrc_smooth_c,
-      ctx.qacc_smooth_c,
-      ctx.qacc_warmstart_c,
-    ],
-  )
-
-
-@event_scope
-def _compact_scatter(m: types.Model, d: types.Data, ctx: NvCompactContext):
-  # scatter results back to full DOF space (inactive frozen to 0) in one launch
-  wp.launch(
-    _scatter_dof_vecs,
-    dim=(d.nworld, m.nv),
-    inputs=[d.dof_cdof, ctx.qacc_c, ctx.qfrc_constraint_c],
-    outputs=[d.qacc, d.qfrc_constraint],
-  )
-
-  # Refresh full d.efc.Ma = M @ qacc. The integrators (Euler/implicit damping) use Ma as
-  # the RHS; the compact solve only populated the compacted Ma, so recompute it in full
-  # space. Inactive DOFs have qacc=0 so their Ma is 0 and they stay frozen.
-  support.mul_m(m, d, d.efc.Ma, d.qacc)
